Moving level loader to common
This required a substantial rework. In the new version, the levels are
static resources and do not get modified by the presense of players.

The camera class changes a bit; it now has hasFocus, setFocus, and
focus. The level is no longer updated and all entities are updated (even
if not on same level).

I also removed the Attack command; it wasn't working and I have quite a
bit of restructuring to do! No point in having dead code.